requirements.txt
crewai==0.130.0
fastapi==0.110.3
uvicorn==0.29.0
groq==0.9.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
langchain-community==0.0.38
pydantic==1.10.13